For consistency, changed all ints to longs when referring to database fields (delta and timestamp)
git-svn-id: http://fbms.googlecode.com/svn/trunk@107 c375ce47-9a4c-63b9-7dad-5ffa0e8ccef7

+++ main/src/cmpt370/fbms/Control.java
@@ -235,17 +235,17 @@ public class Control
 
 	}
 
-	public static void displayRevision(Path file, int timestamp)
+	public static void displayRevision(Path file, long timestamp)
 	{
 
 	}
 
-	public static void displayRevisionChanges(Path file, int timestamp)
+	public static void displayRevisionChanges(Path file, long timestamp)
 	{
 
 	}
 
-	public static void revertRevision(Path file, int timestampt)
+	public static void revertRevision(Path file, long timestampt)
 	{
 
 	}
