+++ main/src/cmpt370/fbms/FileHistory.java
@@ -4,23 +4,23 @@ import java.nio.file.Path;
 
 public class FileHistory
 {
-	public static Path getRevision(Path file, int timestamp)
+	public static Path getRevision(Path file, long timestamp)
 	{
 		return null;
 	}
-	
-	public static void storeRevision(Path diff, int filesize, int delta)
+
+	public static void storeRevision(Path diff, long filesize, long delta)
 	{
-		
+
 	}
-	
-	public static Path obtainRevision(Path file, int timestamp)
+
+	public static Path obtainRevision(Path file, long timestamp)
 	{
 		return null;
 	}
-	
+
 	public static void renameRevision(Path file, String newName)
 	{
-		
+
 	}
 }
